Add the ability to specify number of video frames
This adds the ability to specify a total number of frames to extract
from a video and hash. The reasoning for this is to enable faster
searching where the video being searched doesn't need to have so many
frames extracted, dramatically speeding up the search time.

In addition this adds tests to gurantee this works. All other
functionality is the same.